subprocesses modularized. Crap manager new features
